requirements.txt
Flask==3.0.0
tweepy>=4.14.0
google-auth==2.23.4
google-api-python-client==2.108.0
schedule==1.2.0
gunicorn==21.2.0
pytz==2025.2
python-dotenv==1.1.1